chore(input): 🤖  move inputfield in separate file [#170183355]
